refactor: move conditions to their own subdirectory
