+++ src/Framework/FieldsAPI/Conditions/Condition.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Give\Framework\FieldsAPI\Concerns;
+namespace Give\Framework\FieldsAPI\Conditions;
 
 use JsonSerializable;
